requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paddleocr==2.8.1
opencv-python==4.10.0
pdf2image==1.17.0
requests==2.31.0
pillow==10.4.0